ARCH: arm64: Revert "arm64: Increase the max granular size"
This reverts commit 9730348.

Commit 9730348 ("arm64: Increase the max granular size") increased
the cache line size to 128 to match Cavium ThunderX, apparently for some
performance benefit which could not be confirmed. This change, however,
has an impact on the network packet allocation in certain circumstances,
requiring slightly over a 4K page with a significant performance
degradation. The patch reverts L1_CACHE_SHIFT back to 6 (64-byte cache
line).

Cc: Will Deacon <will.deacon@arm.com>
Cc: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Change-Id: I99610503d9906d0ae6f0f9626aec783538ddfefc